requirements.txt
requests==2.20.1
requests-toolbelt==0.9.1
pycryptodome==3.9.8
antlr4-python3-runtime==4.8